+++ mlflow/__init__.py
@@ -31,6 +31,7 @@ from mlflow.version import VERSION as __version__  # pylint: disable=unused-impo
 from mlflow.utils.logging_utils import _configure_mlflow_loggers
 import mlflow.tracking._model_registry.fluent
 import mlflow.tracking.fluent
+import mlflow.models
 
 # Filter annoying Cython warnings that serve no good purpose, and so before
 # importing other modules.
@@ -151,7 +152,7 @@ delete_experiment = mlflow.tracking.fluent.delete_experiment
 delete_run = mlflow.tracking.fluent.delete_run
 register_model = mlflow.tracking._model_registry.fluent.register_model
 autolog = mlflow.tracking.fluent.autolog
-
+evaluate = mlflow.models.evaluate
 
 run = projects.run
 
@@ -191,4 +192,5 @@ __all__ = [
     "set_registry_uri",
     "list_run_infos",
     "autolog",
+    "evaluate",
 ] + _model_flavors_supported
